+++ lib/fourOhFour.js
@@ -1,6 +1,9 @@
 'use strict'
 
 const FindMyWay = require('find-my-way')
+
+const Reply = require('./reply')
+const Request = require('./request')
 const Context = require('./context')
 const {
   kRoutePrefix,
@@ -27,16 +30,16 @@ const {
   // kFourOhFourContext
 } = require('./symbols.js')
 
-function fourOhFour (fourOhFourFallBack) {
+function fourOhFour (logger, modifyCoreObjects, genReqId) {
   // 404 router, used for handling encapsulated 404 handlers
   const router = FindMyWay({ defaultRoute: fourOhFourFallBack })
 
-  const papi = {
+  const fof = {
     router,
     setNotFoundHandler: setNotFoundHandler
   }
 
-  return papi
+  return fof
 
   function basic404 (req, reply) {
     reply.code(404).send(new Error('Not Found'))
@@ -144,6 +147,28 @@ function fourOhFour (fourOhFourFallBack) {
 
     this[kFourOhFour].router.all(prefix + (prefix.endsWith('/') ? '*' : '/*'), routeHandler, context)
     this[kFourOhFour].router.all(prefix || '/', routeHandler, context)
+  }
+
+  function fourOhFourFallBack (req, res) {
+    // if this happen, we have a very bad bug
+    // we might want to do some hard debugging
+    // here, let's print out as much info as
+    // we can
+    req.id = genReqId(req)
+    req.originalUrl = req.url
+    var childLogger = logger.child({ reqId: req.id })
+    if (modifyCoreObjects) {
+      req.log = res.log = childLogger
+    }
+
+    childLogger.info({ req }, 'incoming request')
+
+    var request = new Request(null, req, null, req.headers, childLogger)
+    var reply = new Reply(res, { onSend: [], onError: [] }, request, childLogger)
+
+    request.log.warn('the default handler for 404 did not catch this, this is likely a fastify bug, please report it')
+    request.log.warn(router.prettyPrint())
+    reply.code(404).send(new Error('Not Found'))
   }
 }
